Split mutation tracker from OptimisticStateTree
